Documentation on `RouterInstructionSyntax`
