+++ kompass-core/src/commonMain/kotlin/io/sellmair/kompass/core/PlainStackInstruction.kt
@@ -1,0 +1,3 @@
+package io.sellmair.kompass.core
+
+typealias PlainStackInstruction<T> = List<RoutingStack.Element<T>>.() -> Iterable<RoutingStack.Element<T>>
